+++ src/components/copy-button/copy-button.test.tsx
@@ -38,14 +38,15 @@ describe('CopyButton', () => {
       className: 'px6 py6 btn btn--purple btn--stroke',
       focusTrapPaused: true,
       block: true,
-      tooltipTheme: 'dark',
+      tooltipColoring: 'dark',
+      tooltipTextSize: 'none',
       children: (
         <span>A custom child in place of the visual button</span>
       ),
       passthroughProps: {
         'data-testid': 'foo'
       }
-    };
+    } as const;
 
     test('renders as expected', () => {
       const { baseElement } = render(<CopyButton {...props} />)