+++ src/components/copy-button/examples/copy-button-c.tsx
@@ -7,7 +7,7 @@ import CopyButton from '../copy-button';
 export default function Example() {
   return (
     <div className="inline-block">
-      <CopyButton themeTooltip="dark" text="Some copy from the custom element">
+      <CopyButton tooltipColoring="dark" tooltipTextSize="xs" text="Some copy from the custom element">
         <button className="w240 btn">Just a custom element</button>
       </CopyButton>
     </div>